+++ catalog/app/components/JsonEditor/Column.tsx
@@ -1,3 +1,4 @@
+import type { JSONType } from 'ajv'
 import cx from 'classnames'
 import * as R from 'ramda'
 import * as React from 'react'
@@ -45,13 +46,13 @@ function getColumnType(
   parent: JsonValue,
 ) {
   const columnSchema = getJsonDictValue(columnPath, jsonDict)
-  if (columnSchema && !parent) return columnSchema.type
+  if (columnSchema && !parent) return columnSchema.type as JSONType
 
   if (Array.isArray(parent)) return 'array'
 
   if (!columnSchema) return 'object'
 
-  return typeof parent
+  return typeof parent as JSONType
 }
 
 const useEmptyColumnStyles = M.makeStyles((t) => ({
@@ -62,8 +63,9 @@ const useEmptyColumnStyles = M.makeStyles((t) => ({
 }))
 
 interface EmptyColumnProps {
-  columnType: 'array' | 'object'
+  columnType?: JSONType
 }
+
 function EmptyColumn({ columnType }: EmptyColumnProps) {
   const classes = useEmptyColumnStyles()
